requirements.txt
# requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# WINDOWS USERS — run this first before pip install:
#   conda install twofish   (or install Microsoft Visual C++ Build Tools)
# Then:
#   pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
#
# KEY FIX vs your previous file:
#   OLD (broken):  hopsworks==4.6.3  +  hsfs==2.2.22  (mismatched major versions)
#   OLD (broken):  hopsworks==3.7.0  +  hsfs==3.7.0   (hsfs 3.7.0 doesn't exist on PyPI)
#   NEW (correct): hopsworks[python]==4.2.10           (hsfs + hsml are BUNDLED inside)
#
# Since Hopsworks 4.x, hsfs and hsml ship INSIDE the hopsworks package.
# Never pin them as separate packages in requirements files.
# ─────────────────────────────────────────────────────────────────────────────

# ── Hopsworks Feature Store + Model Registry (all-in-one) ────────────────────
hopsworks[python]==4.7.2

# ── Core Data Science ────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.2

# ── ML Models ────────────────────────────────────────────────────────────────
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.8

# ── Model Explainability ─────────────────────────────────────────────────────
shap==0.44.1

# ── Data Handling ────────────────────────────────────────────────────────────
pyarrow==15.0.0
joblib==1.3.2

# ── Data Fetching ────────────────────────────────────────────────────────────
requests==2.31.0
python-dotenv==1.0.1

# ── Streamlit & Visualisation ────────────────────────────────────────────────
streamlit==1.31.0
plotly==5.18.0
altair==4.2.2
matplotlib==3.8.2
seaborn==0.13.1

# ── Jupyter / Notebooks ──────────────────────────────────────────────────────
ipykernel==6.28.0
notebook==7.0.6

# ── Utilities ────────────────────────────────────────────────────────────────
python-dateutil==2.9.0.post0
pytz==2024.1
tqdm==4.66.1
pyyaml==6.0.1